More effort towards establishing the resolution of combat using the 3D engine. Specific changes:
- Added a combat turn interface, analogous to the game turn interface, to
ClientApp and AIInterface.  Default AIInterface behavior is to do nothing,
which will mean that units will perform their default missions.

- Added COMBAT_TURN_ORDERS message type, analogous to TURN_ORDERS, to
MessageType.

- Changed ServerCombatStartMessage() to transmit all the UniverseObjects in
the combat system, so that when combats occur as a result of shps visiting
systems for the first time, the attacking plaer can actually see the planets
in the new system.

- Removed cruft from CombatSystem.cpp.  Eventually that whole directory will
go away...

- Implemented the basic framework for handling combat turns among the server,
clients, and AIs.  Currently, no code is in place for actually giving orders
from the clients, or processing those orders once received at the server.

- Now, the COMBAT_START/COMBAT_END pair is only sent out to those players
involved in the combat being resolved.

- Created CombatOrder and CombatOrderSet classes, analogous to the Order class
hierarchy and OrderSet.  The missions used in CombatOrder are based on the
OpenSteer pathfinding prototype.