This commit represents 2 major changes that make the combat turn cycle run much more smoothly:
- The client and server now perform their turn evaluations in parallel, even
when the client is in turn-based mode.  This will make it easier to add a
non-turn-bounded mode, and eliminates the delay at the start of each turn
before anything moved that was there before.

- The client's PathingEngine is now updated at least 100 times per second.
This ensures that it's time steps are close enough together that they make a
smooth curve.  Similarly, the server's PathingEngine is now updated at least
60 times per turn, ensuring that it makes smooth curves too.  This pair of
changes ensures that the two do not get (far) out of sync.